requirements.txt
pytz~=2022.1
datetime~=4.4
telegram~=0.0.1
python-telegram-bot~=13.0
redis==4.2.2
selenium~=4.1.3
python-dotenv~=0.20.0
pip~=22.0.4
#DateTime~=4.4
requests~=2.27.1
